Separated user avatar list from editor header
